+++ node/full/integration_test.go
@@ -1,4 +1,4 @@
-package node
+package full
 
 import (
 	"context"
